isolate nodejs functions so that emulator can consume BotConfigurationBase with no nodejs depdencies
