+++ libraries/botframework-config/src/index.ts
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 export { BotConfiguration } from './botConfiguration';
-export { decryptString, encryptString, generateKey } from './encrypt';
+export { BotConfigurationBase } from './botConfigurationBase';
 export * from './models';
 export * from './schema';
+
