+++ libraries/botframework-config/src/models/appInsightsService.ts
@@ -2,7 +2,6 @@
  * Copyright(c) Microsoft Corporation.All rights reserved.
  * Licensed under the MIT License.
  */
-import { decryptString, encryptString } from '../encrypt';
 import { IAppInsightsService, ServiceTypes } from '../schema';
 import { AzureService } from './azureService';
 
@@ -23,7 +22,7 @@ export class AppInsightsService extends AzureService implements IAppInsightsServ
     }
 
     // encrypt keys in service
-    public encrypt(secret: string): void {
+    public encrypt(secret: string, encryptString: (value: string, secret:string) => string): void {
         if (this.instrumentationKey && this.instrumentationKey.length > 0)
             this.instrumentationKey = encryptString(this.instrumentationKey, secret);
         if (this.apiKeys) {
@@ -34,7 +33,7 @@ export class AppInsightsService extends AzureService implements IAppInsightsServ
     }
 
     // decrypt keys in service
-    public decrypt(secret: string): void {
+    public decrypt(secret: string, decryptString: (value: string, secret:string) => string): void {
         if (this.instrumentationKey && this.instrumentationKey.length > 0)
             this.instrumentationKey = decryptString(this.instrumentationKey, secret);
         if (this.apiKeys) {
