+++ libraries/botframework-config/src/models/azureService.ts
@@ -23,12 +23,12 @@ export class AzureService extends ConnectedService implements IAzureService {
     }
 
     // encrypt keys in service
-    public encrypt(secret: string, iv?: string): void {
+    public encrypt(secret: string, encryptString: (value: string, secret:string) => string): void {
 
     }
 
     // decrypt keys in service
-    public decrypt(secret: string, iv?: string): void {
+    public decrypt(secret: string, decryptString: (value: string, secret:string) => string): void {
 
     }
 }
