+++ libraries/botframework-config/src/models/blobStorageService.ts
@@ -2,7 +2,6 @@
  * Copyright(c) Microsoft Corporation.All rights reserved.
  * Licensed under the MIT License.
  */
-import { decryptString, encryptString } from '../encrypt';
 import { IBlobStorageService, ServiceTypes } from '../schema';
 import { AzureService } from './azureService';
 
@@ -22,13 +21,13 @@ export class BlobStorageService extends AzureService implements IBlobStorageServ
     }
 
     // encrypt keys in service
-    public encrypt(secret: string): void {
+    public encrypt(secret: string, encryptString: (value: string, secret:string) => string): void {
         if (this.connectionString && this.connectionString.length > 0)
             this.connectionString = encryptString(this.connectionString, secret);
     }
 
     // decrypt keys in service
-    public decrypt(secret: string): void {
+    public decrypt(secret: string, decryptString: (value: string, secret:string) => string): void {
         if (this.connectionString && this.connectionString.length > 0)
             this.connectionString = decryptString(this.connectionString, secret);
     }
