+++ libraries/botframework-config/src/models/genericService.ts
@@ -2,7 +2,6 @@
  * Copyright(c) Microsoft Corporation.All rights reserved.
  * Licensed under the MIT License.
  */
-import { decryptString, encryptString } from '../encrypt';
 import { IGenericService, ServiceTypes } from '../schema';
 import { ConnectedService } from './connectedService';
 
@@ -22,7 +21,7 @@ export class GenericService extends ConnectedService implements IGenericService 
     }
 
     // encrypt keys in service
-    public encrypt(secret: string): void {
+    public encrypt(secret: string, encryptString: (value: string, secret:string) => string): void {
         if (this.configuration) {
             for (let prop in this.configuration) {
                 this.configuration[prop] = encryptString(this.configuration[prop], secret);
@@ -31,7 +30,7 @@ export class GenericService extends ConnectedService implements IGenericService 
     }
 
     // decrypt keys in service
-    public decrypt(secret: string): void {
+    public decrypt(secret: string, decryptString: (value: string, secret:string) => string): void {
         if (this.configuration) {
             for (let prop in this.configuration) {
                 this.configuration[prop] = decryptString(this.configuration[prop], secret);
